+++ src/main/kotlin/dsl/Person.kt
@@ -1,0 +1,3 @@
+package dsl
+
+data class Person(val name: String, val company: String?)
