+++ src/main/kotlin/dsl/PersonBuilder.kt
@@ -1,0 +1,22 @@
+package dsl
+
+class PersonBuilder {
+    private lateinit var name: String
+    private var company: String? = null
+
+    fun name(value: String) {
+        name = value
+    }
+
+    fun company(value: String) {
+        company = value
+    }
+
+    fun build(): Person {
+        return Person(name, company)
+    }
+}
+
+fun introduce(block: PersonBuilder.() -> Unit): Person {
+    return PersonBuilder().apply(block).build()
+}
