+++ src/test/kotlin/dsl/DslTest.kt
@@ -1,4 +1,4 @@
-package study
+package dsl
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -42,26 +42,3 @@ class DslTest {
         assertThat(person.company).isEqualTo("카카오")
     }
 }
-
-fun introduce(block: PersonBuilder.() -> Unit): Person {
-    return PersonBuilder().apply(block).build()
-}
-
-class PersonBuilder {
-    private lateinit var name: String
-    private var company: String? = null
-
-    fun name(value: String) {
-        name = value
-    }
-
-    fun company(value: String) {
-        company = value
-    }
-
-    fun build(): Person {
-        return Person(name, company)
-    }
-}
-
-data class Person(val name: String, val company: String?)
